python_react_example copy/backend/requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==3.7.1
azure-core==1.32.0
azure-identity==1.15.0
azure-storage-blob==12.19.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
fastapi==0.104.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
iniconfig==2.0.0
isodate==0.7.2
Mako==1.3.8
MarkupSafe==3.0.2
msal==1.31.1
msal-extensions==1.2.0
packaging==24.2
Pillow==10.1.0
pluggy==1.5.0
portalocker==2.10.1
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.10.1
pytest==7.4.3
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.32.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.24.0
